+++ src/main/java/com/tecfrac/helpdesk/model/ModelGroup.java
@@ -6,8 +6,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
@@ -20,7 +18,7 @@ import javax.persistence.Table;
 @Table(name = "`group`")
 public class ModelGroup {
 
-    public static final int Default = 0;
+    public static final long Default = 0;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
