+++ src/main/java/com/tecfrac/helpdesk/repository/MessageRepository.java
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MessageRepository extends JpaRepository<ModelTicketMessage, Integer> {
+public interface MessageRepository extends JpaRepository<ModelTicketMessage, Long> {
 }
